Issue 280: Set-SafeguardPatch and Install-SafeguardPatch return JSON string but don't throw exceptions for errors
With Safeguard 6.0, some patch precondition warnings and errors may be returned in the response to uploading the patch. This is different from the original behavior in which you manually had to make an additional request to get the patch precondition checks after uploading the patch file. The Cmdlets had been written to first distribute the patch to all cluster nodes before getting and checking the patch preconditions.

These code changes will attempt to inspect the response of the patch file upload for any patch precondition warnings or errors first, before distributing it to other cluster nodes. If the version of Safeguard doesn't return patch precondition checks in the response of the upload, they will be checked after the patch is distributed, as it had been doing.

This also includes a major change to the embedded type that was written in C# to handle the large file upload. With PowerShell now capable of running on Linux and MacOS, it needed to be changed to be compatible with those platforms. Namely, attempting to use `Console` when on Linux didn't work. So it has been changed to utilize the native `Write-Progress` cmdlet of PowerShell. But in order to do that, the embedded code had to be written as a Cmdlet itself.