Issue #545 Identityprovider not beeing used with Connect-Safeguard -Browser
For Issue #545, pass the value of the `-IdentityProvider` parameter to the internal method that launches the web browser for logging in. This will cause Safeguard to default the login page to that provider.

Or, for example, if you specify an External Federation provider, Safeguard will automatically redirect the browser to that provider's login page, including the login hint that came from the `-Username` parameter, if provided.